feat/hub-page-instructor-imporvement-as-props
