+++ src/components/pages/HubView.vue
@@ -10,18 +10,34 @@ import HubStudentQuestionList from '../components/hub-components/HubStudentQuest
 
 // Import store.
 import { useUserDetailsStore } from '../../stores/UserDetailsStore';
+import { useAssessmentsStore } from '../../stores/AssessmentsStore.js';
+import { useUsersStore } from '../../stores/UsersStore.js';
+import { useSkillsStore } from '../../stores/SkillsStore.js';
+import { useInstructorStudentsStore } from '../../stores/InstructorStudentsStore.js';
 
 export default {
     setup() {
         const userDetailsStore = useUserDetailsStore();
+        const assessmentsStore = useAssessmentsStore();
+        const usersStore = useUsersStore();
+        const skillsStore = useSkillsStore();
+        const instructorStudentsStore = useInstructorStudentsStore();
         // Run the GET request.
         userDetailsStore.getUserDetails();
         return {
-            userDetailsStore
+            userDetailsStore,
+            assessmentsStore,
+            usersStore,
+            skillsStore,
+            instructorStudentsStore
         };
     },
     data() {
-        return {};
+        return {
+            assessments: [],
+            questions: [],
+            studentIds: []
+        };
     },
     components: {
         News,
@@ -41,7 +57,130 @@ export default {
             );
         }
     },
-    methods: {}
+    async created() {
+        await this.fetchAssessments();
+        await this.getStudentMCQuestions();
+    },
+    methods: {
+        async fetchAssessments() {
+            // Create the assessments array ---------------------------------
+            // Get unmarked assessments if there no assessment store before
+            await this.assessmentsStore.getAssessments();
+
+            // Get the instructor student list, if not yet loaded.
+            if (
+                this.instructorStudentsStore.instructorStudentsList.length == 0
+            ) {
+                await this.instructorStudentsStore.getInstructorStudentsList();
+            }
+
+            // Just get the students that this instructors teaches.
+            for (
+                let i = 0;
+                i < this.instructorStudentsStore.instructorStudentsList.length;
+                i++
+            ) {
+                if (
+                    this.userDetailsStore.userId ==
+                    this.instructorStudentsStore.instructorStudentsList[i]
+                        .instructor_id
+                ) {
+                    this.studentIds.push(
+                        this.instructorStudentsStore.instructorStudentsList[i]
+                            .student_id
+                    );
+                }
+            }
+            // Get the assessments for those students.
+            for (let i = 0; i < this.assessmentsStore.assessments.length; i++) {
+                for (let j = 0; j < this.studentIds.length; j++) {
+                    if (
+                        this.assessmentsStore.assessments[i].student_id ==
+                        this.studentIds[j]
+                    ) {
+                        this.assessments.push(
+                            this.assessmentsStore.assessments[i]
+                        );
+                    }
+                }
+            }
+
+            // Date.
+            for (let i = 0; i < this.assessments.length; i++) {
+                let date = new Date(this.assessments[i].date).toDateString();
+                this.assessments[i].date = date;
+            }
+
+            // Get users.
+            if (this.usersStore.users.length == 0) {
+                await this.usersStore.getUsers();
+            }
+            // Add the student name.
+            for (let i = 0; i < this.assessments.length; i++) {
+                for (let j = 0; j < this.usersStore.users.length; j++) {
+                    if (
+                        this.assessments[i].student_id ==
+                        this.usersStore.users[j].id
+                    ) {
+                        this.assessments[i].studentUsername =
+                            this.usersStore.users[j].username;
+                    }
+                }
+            }
+
+            // Get skills.
+            if (this.skillsStore.skillsList.length == 0) {
+                await this.skillsStore.getSkillsList();
+            }
+            // Add the skill name.
+            for (let i = 0; i < this.assessments.length; i++) {
+                for (let j = 0; j < this.skillsStore.skillsList.length; j++) {
+                    if (
+                        this.assessments[i].skill_id ==
+                        this.skillsStore.skillsList[j].id
+                    ) {
+                        this.assessments[i].skillName =
+                            this.skillsStore.skillsList[j].name;
+                    }
+                }
+            }
+        },
+        formatDate(unformattedDate) {
+            // Prep the date and time data ---------------
+            // Split timestamp into [ Y, M, D, h, m, s ]
+            var date = unformattedDate.replace('T', ' ');
+            date = date.replace('Z', ' ');
+            let newDate = date.split(/[- :]/);
+            // Apply each element to the Date function
+            var finalDate = new Date(
+                Date.UTC(
+                    newDate[0],
+                    newDate[1] - 1,
+                    newDate[2],
+                    newDate[3],
+                    newDate[4],
+                    newDate[5]
+                )
+            );
+            var options = {
+                weekday: 'long',
+                year: 'numeric',
+                month: 'long',
+                day: 'numeric',
+                hour: 'numeric',
+                minute: 'numeric'
+            };
+            finalDate = finalDate.toLocaleDateString('en-US', options);
+            return finalDate;
+        },
+        async getStudentMCQuestions() {
+            const result = await fetch(
+                '/questions/student-mc-questions/full-data-list'
+            );
+            const data = await result.json();
+            this.questions = data;
+        }
+    }
 };
 </script>
 
@@ -52,6 +191,11 @@ export default {
             <div
                 class="col-lg-4 col-md-6 mb-2"
                 v-if="userDetailsStore.role != 'editor'"
+                :class="{
+                    'd-none':
+                        userDetailsStore.role == 'instructor' &&
+                        assessments.length === 0
+                }"
             >
                 <div class="hub-component">
                     <StudentProgress
@@ -60,6 +204,7 @@ export default {
                     />
                     <MarkAssessment
                         v-else-if="userDetailsStore.role == 'instructor'"
+                        :assessments="assessments"
                     />
                 </div>
             </div>
@@ -67,6 +212,11 @@ export default {
             <div
                 class="col-lg-4 col-md-6 mb-2"
                 v-if="userDetailsStore.role != 'editor'"
+                :class="{
+                    'd-none':
+                        userDetailsStore.role == 'instructor' &&
+                        questions.length === 0
+                }"
             >
                 <div class="hub-component">
                     <LastVisitedSkills
@@ -76,6 +226,7 @@ export default {
                     <!-- Student Added Questions List -->
                     <HubStudentQuestionList
                         v-else-if="userDetailsStore.role == 'instructor'"
+                        :questions="questions"
                     />
                 </div>
             </div>
